requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
anthropic>=0.18.0
flask>=3.0.0
markdown>=3.5.0
